重构
将数据库的相关操作提取出来一个类
